+++ src/resources/views/channel/show.blade.php
@@ -5,10 +5,46 @@
 @stop
 
 @push('page-actions')
-    <x-appshell::standard-actions :model="$channel" route="vanilo.admin.channel" :name="$channel->name" />
+    <x-appshell::standard-actions :model="$channel" route="vanilo.admin.channel" :name="$channel->name"/>
 @endpush
 
 @section('content')
+    <div class="row mb-4">
+        <div class="col">
+            <x-appshell::card-with-icon icon="channel" type="info">
+                {{ $channel->name }}
+                <x-slot:subtitle>
+                    {{ $channel->slug }} | {{ $channel->domain ? __('domain: :domain', ['domain' => $channel->domain]) : __('No domain assigned') }}
+                </x-slot:subtitle>
+            </x-appshell::card-with-icon>
+        </div>
+
+            <div class="col">
+                <x-appshell::card-with-icon icon="globe" type="success">
+                    {{ $country?->name ?? __('No country assigned') }}
+
+                    <x-slot:subtitle>
+                        {{ ($channel->currency ?? config('vanilo.foundation.currency.code')) }} | {{ ($channel->language ?? substr(app()->getLocale(), 0, 2)) }}
+                    </x-slot:subtitle>
+                </x-appshell::card-with-icon>
+            </div>
+
+        <div class="col">
+            <x-appshell::card-with-icon icon="bag">
+                {{ __(':num purchases', ['num' => $orderCount]) }}
+
+                <x-slot:subtitle>
+                    @foreach ($orderTotalsPerCurrency as $currency => $sum)
+                        {{ $currency }}: {{ $sum }}
+                        @if (!$loop->last)
+                            |
+                        @endif
+                    @endforeach
+                </x-slot:subtitle>
+            </x-appshell::card-with-icon>
+        </div>
+    </div>
+
     <x-appshell::card>
         <x-slot:title>{{ __('Assigned Products') }}</x-slot:title>
         <x-slot:actions>